+++ src/components/mypage/MyBagList.tsx
@@ -89,7 +89,9 @@ export default function MyBagList() {
                     {datas === undefined ? (
                         ''
                     ) : datas.length === 0 ? (
-                        <span className='flex justify-center'>Loading...</span>
+                        <span className='flex justify-center'>
+                            여행 일정이 없습니다.
+                        </span>
                     ) : (
                         datas.map((data: ScheduleProps) => (
                             <div
