+++ src/components/schedulemain/SelectedPlaces.tsx
@@ -16,7 +16,7 @@ export default function SelectedPlaces({
     };
 
     return (
-        <div className='flex flex-col bg-brightgrey rounded-md px-[33px] py-2.5 mx-4'>
+        <div className='flex flex-col bg-brightgrey rounded-md px-[33px] py-2.5 mx-4 mt-4'>
             <div className='py-2.5'>여행 일정 등록하기</div>
             <div className='flex py-2.5'>
                 <div className='flex text-center px-3 py-2 mr-2 rounded-full border border-grey text-[12px] text-darkgrey'>
