+++ src/components/summary/MyTravel.tsx
@@ -90,7 +90,9 @@ export default function MyTravel() {
                     {datas === undefined ? (
                         ''
                     ) : datas.length === 0 ? (
-                        <span className='flex justify-center'>Loading...</span>
+                        <span className='flex justify-center'>
+                            여행 일정이 없습니다.
+                        </span>
                     ) : (
                         datas.map((data: ScheduleProps) => (
                             <div
